refactor: move pseudoprior stuff to its own subpackage
